+++ tests/testthat/test-all-major-funs.R
@@ -48,13 +48,14 @@ mp <- structure(list(a = 1:2, geometry = structure(list(structure(c(0,
 
 
 r2 <- setExtent(raster(volcano), extent(-2, 10, -5, 14))
-
+library(sf)
 test_that("extract of sf works", {
   cellnumbers(r, psf[c(1, 10), ]) %>% expect_named(c("object_", "cell_")) 
-  ll <- extract(r, psf) %>% expect_length(24) %>% lengths()
-  expect_that(sum(ll), equals(5307))
-  #extract(r, as(psf, "Spatial")) %>% lengths()
+  ll <- extract(r, psf[1:5, ]) %>% expect_length(5) %>% lengths()
+  expect_that(sum(ll), equals(1232))
+  e <- extract(r, as(psf, "Spatial")[1:6, ]) %>% lengths()
+  expect_equal(e, c(256L, 256L, 256L, 208L, 256L, 256L))
   ## awaiting fix in spbabel https://github.com/r-gris/tabularaster/issues/8
-  cellnumbers(r2, mp)
+  cellnumbers(r2, mp) %>% expect_named(c("object_", "cell_"))
 })