+++ NEWS.md
@@ -1,4 +1,4 @@
-# tabularaster dev
+# tabularaster 0.6.0
 
 * Function `cellnumbers` now provides only basic cell lookup, the weights and other
  options available in raster functions are ignored.